pull prompt into CLI
